// neuron_viewer/src/client/models/GroupId.ts
// Auto-generated code. Do not edit! See neuron_explainer/activation_server/README.md to learn how to regenerate it.

/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Identifiers for groups in multi-top-k requests.
 */
export enum GroupId {
  ACT_TIMES_GRAD = "act_times_grad",
  ACTIVATION = "activation",
  DIRECT_WRITE_TO_GRAD = "direct_write_to_grad",
  DIRECTION_WRITE = "direction_write",
  LOGITS = "logits",
  MLP_LAYER_WRITE = "mlp_layer_write",
  SINGLETON = "singleton",
  TOKEN_WRITE = "token_write",
  TOKEN_READ = "token_read",
  WRITE_NORM = "write_norm",
  TOKEN_PAIR_ATTRIBUTION = "token_pair_attribution",
}